Organise tasks into sections
Also add a `getSection` helper to get a section by its class. Some
“Sections” (such as Check Your Answers) are both Sections and
Tasks
